requirements.txt
streamlit 
openai-whisper
torch
python-dotenv
llama-index
groq
